requirements.txt
pbr
WebOb
tabulate
click
Flask
six
pyyaml